Win32 changes
git-svn-id: https://svn.ntop.org/svn/ntop/trunk/nDPI@6923 a02cd0c1-8e76-42e1-a119-56f9641475e2

+++ src/lib/ndpi_cache.c
@@ -34,7 +34,7 @@ static u_int8_t traceLRU = 0;
 
 static u_int32_t get_now(void) {
 #ifndef __KERNEL__
-  return(time(NULL));
+  return((u_int32_t)time(NULL));
 #else
   return(jiffies);
 #endif
